+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [7.0.0] - 2024-12-18
 - Rewritten Pynamo & Mongo abstraction to support all attributes and functionality:
   - `modular_sdk.models.pynamongo.models.Model` must be used instead of old BaseModel (see docs inside `modular_sdk.models.pynamongo.__init__`)
-  - `get_json()` is not deprecated and will be removed. Use pynamongo.convertors.instance_to_dict or write your serializer
+  - `get_json()` is removed. Use pynamongo.convertors.instance_to_dict or write your serializer
   - `mongo_id` attribute was replaced with `__mongo_id__`
 - Updated `pymongo` version to `pymongo~=4.10.1`
 - Removed `dynamodb-json~=1.4.2` package